refactor: 修复 test 用例，修改为 --reporter min
