+++ packages/matman-cli/test/mocha.opts
@@ -1,3 +1,3 @@
 test/scripts
---reporter spec
+--reporter min
 --recursive
