+++ packages/matman/test/scripts/model/ScreenshotConfig.test.js
@@ -7,22 +7,21 @@ import { findMatmanConfig } from '../../../src/util';
 describe('check model/ScreenshotConfig.js', () => {
     describe('check constructor(opts, basePath, tag)', () => {
         const MATMAN_ROOT_PATH = path.join(__dirname, '../../data/hi-demo/demo_01');
-        const caseModuleFilePath = path.join(MATMAN_ROOT_PATH, './src/page_baidu_index/cases/basic-check.js');
+        const caseModuleFilePath = path.join(MATMAN_ROOT_PATH, './case_modules/page_baidu_index/basic-check.js');
         const matmanConfig = findMatmanConfig(caseModuleFilePath,{
-            rootPath: MATMAN_ROOT_PATH,
-            caseModulesPath: path.join(MATMAN_ROOT_PATH, './src')
+            rootPath: MATMAN_ROOT_PATH
         });
 
         it('if opts is Boolean and tag is undefined', () => {
             expect(new ScreenshotConfig(matmanConfig, true, caseModuleFilePath)).to.eql({
-                path: path.join(matmanConfig.screenshotPath, './page_baidu_index/cases/basic-check_js/basic-check_js.png'),
+                path: path.join(matmanConfig.screenshotPath, './page_baidu_index/basic-check_js/basic-check_js.png'),
                 tag: undefined
             });
         });
 
         it('if opts is Boolean and tag is "mytag"', () => {
             expect(new ScreenshotConfig(matmanConfig, true, caseModuleFilePath, 'mytag')).to.eql({
-                path: path.join(matmanConfig.screenshotPath, './page_baidu_index/cases/basic-check_js/basic-check_js_mytag.png'),
+                path: path.join(matmanConfig.screenshotPath, './page_baidu_index/basic-check_js/basic-check_js_mytag.png'),
                 tag: 'mytag'
             });
         });
@@ -160,14 +159,13 @@ describe('check model/ScreenshotConfig.js', () => {
 
     describe('check getPathWithId(id)', () => {
         const MATMAN_ROOT_PATH = path.join(__dirname, '../../data/hi-demo/demo_01');
-        const caseModuleFilePath = path.join(MATMAN_ROOT_PATH, './src/page_baidu_index/cases/basic-check.js');
+        const caseModuleFilePath = path.join(MATMAN_ROOT_PATH, './case_modules/page_baidu_index/basic-check.js');
         const matmanConfig = findMatmanConfig(caseModuleFilePath,{
-            rootPath: MATMAN_ROOT_PATH,
-            caseModulesPath: path.join(MATMAN_ROOT_PATH, './src')
+            rootPath: MATMAN_ROOT_PATH
         });
 
         it('if opts is Boolean and tag is "mytag"', () => {
-            expect(new ScreenshotConfig(matmanConfig, true, caseModuleFilePath, 'mytag').getPathWithId(1024)).to.equal(path.join(matmanConfig.screenshotPath, './page_baidu_index/cases/basic-check_js/basic-check_js_mytag_1024.png'));
+            expect(new ScreenshotConfig(matmanConfig, true, caseModuleFilePath, 'mytag').getPathWithId(1024)).to.equal(path.join(matmanConfig.screenshotPath, './page_baidu_index/basic-check_js/basic-check_js_mytag_1024.png'));
         });
 
         it('if opts is String(relative path) and tag is "mytag"', () => {
