+++ test/scripts/demo_02/search-check.test.js
@@ -1,6 +1,6 @@
 const expect = require('chai').expect;
 
-const checkPage = require('../../../packages/matman/test/data/hi-demo/demo_02/src/page_baidu_index/cases/search-check');
+const checkPage = require('../../../packages/matman/test/data/hi-demo/demo_02/case_modules/page_baidu_index/search-check');
 
 describe('百度首页：搜索', function () {
     this.timeout(30000);
